% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{amicusCorpus}
\alias{amicus}
\alias{amicusCorpus}
\alias{amicusTexts}
\title{amicus curiae briefs from Bakke (1978) and Bollinger (2008)}
\format{\code{amicusCorpus} is corpus with 100 texts, and \code{docvar} set for petitioner/respondent
\cr \cr \code{amicusTexts} is the same set of texts as a named character vector}
\description{
Texts of amicus curiae briefs labelled as being either pro-petitioner or pro-respondent
in US Supreme court cases on affirmative action,
Bakke (1978) and Bollinger (2008), taken from Evans et al (2007).
}
\references{
Evans, Michael, et al. 2007. "Recounting the courts? Applying automated content analysis to enhance empirical legal research." Journal of Empirical Legal Studies 4(4): 1007-1039.
}

