% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_amicus}
\alias{data_corpus_amicus}
\title{amicus curiae briefs from Bakke (1978) and Bollinger (2008)}
\format{\code{amicusCorpus} is corpus with 100 texts, and \code{docvar} set
  for petitioner/respondent}
\usage{
data_corpus_amicus
}
\description{
Texts of amicus curiae briefs labelled as being either pro-petitioner or
pro-respondent in US Supreme court cases on affirmative action, Bakke (1978)
and Bollinger (2008), taken from Evans et al (2007).
}
\references{
Evans, Michael, et al. 2007. "Recounting the courts? Applying
  automated content analysis to enhance empirical legal research." 
  \emph{Journal of Empirical Legal Studies} 4(4): 1007-1039.
}
\keyword{data}
