% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_immigrationnews}
\alias{data_corpus_immigrationnews}
\title{UK news articles (2,833) from 2014 that mention immigration}
\format{The corpus consists of 2,883 documents and includes the following document variables: \describe{
 \item{paperName}{Character indicating the newspaper.}
 \item{day}{Number of days from Jan 1st 2014.}
 \item{id}{Unique idenifier.}
 }}
\usage{
data_corpus_immigrationnews
}
\description{
A corpus of articles form the UK press in 2014 that mention immigration.
}
\references{
Nulty, P. & Poletti, M. (2014). "The Immigration Issue in the UK in the 2014 
  EU Elections: Text Mining the Public Debate." Presentation at LSE Text Mining Conference 2014.
}
\keyword{datasets}
