% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_udhr}
\alias{data_corpus_udhr}
\title{Universal Declaration of Human Rights}
\format{
The corpus includes the following document variables: \describe{
\item{langauge}{specifies the language of the texts}
\item{LangID}{short identifier of the language}
}
}
\usage{
data_corpus_udhr
}
\description{
A corpus object containing the Universal Declaration of Human Rights in 280
languages.  The files where downloaded from
\url{http://research.ics.aalto.fi/cog/data/udhr/}.  These have been converted
into plain text format by Vatanen, Väyrynen, and Virpioja (2010).  At the
source, there are additional versions in pdf but these were unavailable in
plain text format.
}
\references{
Tommi Vatanen, Jaakko J. Väyrynen, and Sami Virpioja (2010)
Language identification of short text segments with n-gram models. In
\emph{Proceedings of the Seventh International Conference on Language Resources
and Evaluation} (LREC'10), pp. 3423-3430. European Language Resources
Association (ELRA).
}
\keyword{datasets}
