% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_ungd2017}
\alias{data_corpus_ungd2017}
\title{UN General Debate speeches, 2017}
\format{The corpus includes the following document variables: \describe{
  \item{country_iso}{ISO3c country code}
  \item{un_session}{UN session}
  \item{year}{4-digit year}
  \item{country}{country code}
  \item{continent}{continent of the country. Note that the speech delivered on behalf of the 
  European Union is coded as "Europe".}
  \item{gdp}{GDP (in US-$, 2017), derived from World Bank. C
  Variable contains missing values for some countries.}
  \item{gdp_per_capita}{GDP per capita (in US-$, 2017), derived from World Bank. 
  Variable contains missing values for some countries.}
  }}
\source{
Mikhaylov, M., Baturo, A., & Dasandi, N. (2017). United Nations General Debate Corpus. 
  Harvard Dataverse, V4. URL: \url{https://doi.org/10.7910/DVN/0TJX8Y}.
}
\usage{
data_corpus_ungd2017
}
\description{
A corpus of 196 speeches from the 2017 UN General Debate. 
The raw corpus with all speeches since 1970 is available at: \url{https://doi.org/10.7910/DVN/0TJX8Y}. 
The economic data for 2017 (GDP and GDP per capita) are downloaded from the World Bank website.
}
\references{
Mikhaylov, M., Baturo, A., & Dasandi, N. (2017). United Nations General Debate Corpus. 
  Harvard Dataverse, V4. URL: \url{https://doi.org/10.7910/DVN/0TJX8Y}.
  
  Baturo, A., Dasandi, N., & Mikhaylov, S. (2017). \href{10.1177/2053168017712821}{Understanding State Preferences With Text As Data: 
  Introducing the UN General Debate Corpus}. \emph{Research and Politics} 4(2): 1--9.
}
\keyword{datasets}
