% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{data_download}
\alias{data_download}
\title{Donlowad RDS file from online locations}
\usage{
data_download(name = NULL, url = NULL, cache = TRUE, ...)
}
\arguments{
\item{name}{name of a RDS file to download}

\item{url}{user-specified location of the RDS file to download}

\item{cache}{if \code{TRUE}, save downloaded file to temporary folder}

\item{...}{extra argument passed to \code{download.file}}
}
\value{
an unserialized R object
}
\description{
Donlowad RDS file from online locations
}
\details{
Datasets currently available by \code{name} are: \describe{
  \item{\code{data_corpus_guardian}}{6,000 Guradian newspaper articles in
  politics, economy, society and international sections from 2012 to 2016.}
  \item{\code{data_corpus_foreignaffairscommittee}}{Transcripts of all the 
  speeches at Japan's Committee on Foreign Affairs and Defense of the lower 
  house (Shugiin) from 1947 to 2017}.}
}
\examples{
\\donotrun{
corp <- data_download('data_corpus_guardian')
corp <- data_download('data_corpus_foreignaffairscommittee')
}
}
