% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{iebudgetsCorpus}
\alias{ie2010Corpus}
\alias{ieDocvars}
\alias{ieTexts}
\alias{iebudgets}
\alias{iebudgetsCorpus}
\title{Irish budget speeches}
\format{\code{iebudgets} is the corpus object for the 2008-2012 budget
speeches, with document-level variables for year, debate, serial number,
first and last name of the speaker, and the speaker's party. \cr \cr
\code{ieTexts} is a character vector of five budget speeches for
experimenting with corpus construction.\cr \cr \code{ieDocvars} is a data
frame of document-level variables for experimenting with corpus construction.
\cr \cr \code{iebudgets} is the defunct name for the \code{iebudgetsCorpus} object --
use that instead
\cr \cr \code{ie2010Corpus} is the subset for the 2010 budget speeches from
Lowe and Benoit (2013)}
\description{
Speeches and document-level variables from Irish budget debates
  held annually, for the years 2008-2012.
}
\examples{
# load the Irish budget speeches
data(iebudgetsCorpus)
ie2010 <- subset(iebudgets, year==2010)  # just the 2010 speeches
summary(ie2010)
summary(subset(iebudgets, no=="02"))  # just the Finance Minister (always first)
                                     # note that "no" is a factor, not integer

data(ieTexts)
# create a corpus (just five speeches)
mycorpus <- corpus(ieTexts, docvars=ieDocvars)
summary(mycorpus)
}
\references{
Alexander Herzog and Kenneth Benoit (2013)

"\href{http://papers.ssrn.com/sol3/Delivery.cfm/SSRN_ID2281311_code1869090.pdf?abstractid=2225069&mirid=1}{The
Most Unkindest Cuts: Government Cohesion and Economic Crisis.}" Paper
prepared for presentation at the 3rd Annual General Conference of the
European Political Science Association, 20-22 June 2013, Barcelona. \cr \cr
Lowe, Will and Kenneth Benoit. 2013.

"\href{http://www.kenbenoit.net/pdfs/Political%20Analysis-2013-Lowe-298-313.pdf}{Validating
Estimates of Latent Traits From Textual Data Using Human Judgment as a
Benchmark.}" \emph{Political Analysis} 21: 298–313.
}

