% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{movies}
\alias{movies}
\title{movie reviews from Pang, Lee, and Vaithyanathan (2002)}
\description{
A corpus object containing 2000 movie reviews classified by positive or negative sentiment
}
\references{
\url{http://dl.acm.org/citation.cfm?id=1118704}
}

