% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{moviesCorpus}
\alias{movies}
\alias{moviesCorpus}
\title{movie reviews from Pang, Lee, and Vaithyanathan (2002)}
\description{
A corpus object containing 2000 movie reviews classified by positive or negative sentiment
}
\references{
\url{http://dl.acm.org/citation.cfm?id=1118704}
}

