% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{quanteda.corpora-package}
\alias{quanteda.corpora}
\alias{quanteda.corpora-package}
\title{An R package containing texts and corpora for \link[quanteda]{quanteda}.}
\description{
A set of texts and corpus objects for use with the quanteda R package.
}
\seealso{
\link[quanteda]{quanteda}
}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk}

Other contributors:
\itemize{
  \item Kohei Watanabe \email{watanabe.kohei@gmail.com} [contributor]
}

}
