% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{SOTUCorpus}
\alias{SOTUCorpus}
\alias{sotuCorp}
\title{State of the Union addresses from 1985 to 2014}
\description{
A corpus object 30 US State-of-the-Union addresses, 1985 to 2014
}

