% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_immigrationnews}
\alias{data_corpus_immigrationnews}
\title{UK news articles (2,833) from 2014 that mention immigration}
\format{\code{immigNewsCorpus} is corpus with 2833 texts, and \code{docvars}
  set for \code{paperName}, \code{day}, and \code{id}. \code{day} is the
  number of days from Jan 1st 2014.}
\usage{
data_corpus_immigrationnews
}
\description{
A corpus of articles form the UK press in 2014 that mention the 
  immigration. See working paper: 
  "\href{https://github.com/pnulty/LSEtext/blob/master/NultyPoletti.pdf}{The
  Immigration Issue in the European Electoral Campaign in the UK: Text-Mining
  Public Debate from Newspapers and Social Media}"
}
\keyword{datasets}
