% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{data_corpus_movies}
\alias{data_corpus_movies}
\title{movie reviews from Pang, Lee, and Vaithyanathan (2002)}
\format{An object of class \code{corpus} (inherits from \code{list}) of length 4.}
\usage{
data_corpus_movies
}
\description{
A corpus object containing 2000 movie reviews classified by positive or negative sentiment
}
\references{
\url{http://dl.acm.org/citation.cfm?id=1118704}
}
\keyword{data}
