% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_ukmanifestos}
\alias{data_corpus_ukmanifestos}
\title{UK political party manifestos, 1945--2005}
\format{The corpus includes the following document variables: \describe{
  \item{Country}{Country name (UK).}
  \item{Type}{Character indicating the type of election: national (\code{natl}) or regional (\code{regl}.}
  \item{Language}{Language (\code{en}).}
  \item{Party}{A character abbreviation of the political party.}
  }}
\usage{
data_corpus_ukmanifestos
}
\description{
A corpus object containing 105 UK Manifestos from 1945--2005, with party and year attributes
}
\keyword{datasets}
