% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_ukmanifestos}
\alias{data_corpus_ukmanifestos}
\title{UK political party manifestos, 1945-2005}
\format{An object of class \code{corpus} (inherits from \code{list}) of length 4.}
\usage{
data_corpus_ukmanifestos
}
\description{
A corpus object containing 105 UK Manifestos from 1945-2005, with party and year attributes
}
\references{
As used in Laver, Michael. 1998a. \'Party Policy in Britain, 1997: Results from an Expert Survey.\' Political Studies 46: 336–47.
}
\keyword{datasets}
