% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{exampleString}
\alias{exampleString}
\title{A paragraph of text for testing various text-based functions}
\format{character vector with one element}
\description{
This is a long paragraph (2,914 characters) of text taken from an Irish budget speech by Joe Higgins
}
\examples{
data(exampleString)
clean(exampleString)
}

