% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{iebudgetsCorpus}
\alias{iebudgetsCorpus}
\title{Irish budget speeches}
\description{
Speeches and document-level variables from Irish budget debates 
  held annually, for the years 2008-2012.
}
\examples{
# load the Irish budget speeches
data(iebudgetsCorpus)
ie2010 <- subset(iebudgets, year == 2010)  # just the 2010 speeches
summary(ie2010)
summary(subset(iebudgets, no == "02"))  # just the Finance Minister (always first)
                                        # note that "no" is a factor, not integer

data(ieTexts)
# create a corpus (just five speeches)
mycorpus <- corpus(ieTexts, docvars = ieDocvars)
summary(mycorpus)
}
\references{
Alexander Herzog and Kenneth Benoit. 2015.  "\href{http://www.kenbenoit.net/pdfs/Herzog_Benoit_JOP_2015.pdf}{The Most
  Unkindest Cuts: Speaker Selection and Expressed Government Dissent During
  Economic Crisis}". 
  \emph{Journal of Politics} 77(4, October): 1157-1175.
  
  Lowe, Will and Kenneth Benoit. 2013. 
  "\href{http://www.kenbenoit.net/pdfs/Political\%20Analysis-2013-Lowe-298-313.pdf}{Validating
   Estimates of Latent Traits From Textual Data Using Human Judgment as a 
  Benchmark.}" \emph{Political Analysis} 21: 298-313.
}

