% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{package}
\name{quantedaData}
\alias{quantedaData}
\alias{quantedaData-package}
\title{An R package containing texts and corpora for \link[quanteda]{quanteda}.}
\description{
A set of texts and corpus objects for use with the quanteda R package.
}
\seealso{
\link[quanteda]{quanteda}
}
\author{
Ken Benoit and Paul Nulty
}
