% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{ukManifestosCorpus}
\alias{ukManifestosCorpus}
\title{A corpus object containing 105 UK Manifestos}
\description{
A corpus object containing 105 UK Manifestos from 1945-2005, with party and year attributes
}

